[config] improve vm_config
- Testing relied on receiving the whole NodeConfig, now it only needs
VMConfig
- Default on disk did not match default in vm_config, fixed
- Moved helpers that are only used in one place into that same file
- Simplified random node generation now that VM doesn't use it for
testing
