+++ config/src/config/vm_config.rs
@@ -1,10 +1,9 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::utils::{deserialize_whitelist, serialize_whitelist};
 use libra_types::transaction::SCRIPT_HASH_LENGTH;
-use serde::{Deserialize, Serialize};
-use std::collections::HashSet;
+use serde::{Deserialize, Deserializer, Serialize, Serializer};
+use std::{collections::HashSet, hash::BuildHasher};
 
 /// Holds the VM configuration, currently this is only the publishing options for scripts and
 /// modules, but in the future this may need to be expanded to hold more information.
@@ -16,8 +15,30 @@ pub struct VMConfig {
 
 impl Default for VMConfig {
     fn default() -> VMConfig {
+        let whitelist = vec![
+            "6aabc87f543f85e10216432d02b0251297d4c7723e906de481dfa04b057c2371",
+            "1cf66b5f5c911e80dad222b8ee8dfe3ad4830f75bb412ba12ea8e429203d9c83",
+            "a2180395d1632a0793f34e8a8a6be20b3b03bdceee35affe8c751fc8467b73a4",
+            "5ee07d4ac1ecf88f1b41c2c458f15699fe9d811c61563338253b3807b75c04c1",
+        ]
+        .iter()
+        .map(|s| string_to_script_hash(s))
+        .collect();
+
         VMConfig {
-            publishing_options: VMPublishingOption::Open,
+            publishing_options: VMPublishingOption::Locked(whitelist),
+        }
+    }
+}
+
+impl VMConfig {
+    /// Creates a new `VMConfig` where the whitelist is empty. This should only be used for testing.
+    #[allow(non_snake_case)]
+    #[doc(hidden)]
+    #[cfg(any(test, feature = "fuzzing"))]
+    pub fn empty_whitelist_FOR_TESTING() -> Self {
+        VMConfig {
+            publishing_options: VMPublishingOption::Locked(HashSet::new()),
         }
     }
 }
@@ -57,14 +78,37 @@ impl VMPublishingOption {
     }
 }
 
-impl VMConfig {
-    /// Creates a new `VMConfig` where the whitelist is empty. This should only be used for testing.
-    #[allow(non_snake_case)]
-    #[doc(hidden)]
-    #[cfg(any(test, feature = "fuzzing"))]
-    pub fn empty_whitelist_FOR_TESTING() -> Self {
-        VMConfig {
-            publishing_options: VMPublishingOption::Locked(HashSet::new()),
-        }
-    }
+fn string_to_script_hash(input: &str) -> [u8; SCRIPT_HASH_LENGTH] {
+    let mut hash = [0u8; SCRIPT_HASH_LENGTH];
+    let decoded_hash =
+        hex::decode(input).expect("Unable to decode script hash from configuration file.");
+    assert_eq!(decoded_hash.len(), SCRIPT_HASH_LENGTH);
+    hash.copy_from_slice(decoded_hash.as_slice());
+    hash
+}
+
+fn deserialize_whitelist<'de, D>(
+    deserializer: D,
+) -> Result<HashSet<[u8; SCRIPT_HASH_LENGTH]>, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    let whitelisted_scripts: Vec<String> = Deserialize::deserialize(deserializer)?;
+    let whitelist = whitelisted_scripts
+        .iter()
+        .map(|s| string_to_script_hash(s))
+        .collect();
+    Ok(whitelist)
+}
+
+fn serialize_whitelist<S, H>(
+    whitelist: &HashSet<[u8; SCRIPT_HASH_LENGTH], H>,
+    serializer: S,
+) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+    H: BuildHasher,
+{
+    let encoded_whitelist: Vec<String> = whitelist.iter().map(hex::encode).collect();
+    encoded_whitelist.serialize(serializer)
 }
